Add fallbacks to elf code section disassembly. (Closes: #879003)
Because objdump --line-numbers can be extremely slow, it can be worth
skipping it, and just get a disassembly (especially when the files don't
contain line numbers debug info, where ironically, objdump is slow with
--line-numbers).

The --exclude-command command line option allows to skip some commands,
and one can use that to skip the command that uses --line-numbers, but
in that case, no difference is shown for the code sections.

This change adds a fallback in that case, where objdump is called
without the --line-numbers command, making disassembly faster.

But that can still be too time consuming (especially because the output
can be large, and diffing that can be a long process, so allow to
exclude /that/ too, and fallback to an hexdump in that case.