Fix tests under colord >= 1.4.3. (Closes: #908900)
The `cd-iccdump` command has no --version command and, whilst the `colord`
binary does, it relies on the daemon running to return any output.  Therefore
we hackily compare the output (via the line length) of colord 1.4.3:

    "  Profile ID    = 0477fa4bb5ae5ae9a778f5cd72eb45a4"

... versus, for example, colord 1.3.3:

    " Profile ID    = 0x0477fa4b"

We don't massage the output (say, in an `Iccdump.filter` method) as it would
remove the accuracy and, unfortunately, colord 1.4.3 also removes a
somewhat-arbitrary newline too.