Add --list-missing-tools
While creating a Flatpak package for Diffoscope, I found myself wanting
to compare the list reported by --list-tools to what was available in
the sandbox (so that I could add interesting external tools). Since
Diffoscope already has all the machinery to find external tools –
including logic to prefix "g" for GNU tools on non-GNU systems – it is
best placed to report the list of missing tools.