Add a python_requires to setup.py
This means `pip install diffoscope` will not try to download new source
distributions that then fail to build, if you run pip with an unsupported
Python version.  (Unfortunately pip will still attempt to download older
releases of diffoscope that did not have python_requires in the metadata.
Still, it'll be useful when you eventually drop support for Python 3.5 and
start requiring newer Pythons.)