Add support for comparing WebAssembly modules. (MR: !17)
This only defines one WebAssembly dumper in wasm.py ideally, I’d run
`wasm2wat` and if that does not show the difference, run `wasm-objdump`.
Which is a general pattern that should be abstracted in the `Difference`
class somewhere. Might do if the itch becomes large enough, but not
right now.

requires `wabt`, which is not packaged for Debian yet, but is available
elsewhere, e.g. in Arch and nixpkgs.

Signed-off-by: Chris Lamb <lamby@debian.org>